+++ toda/Algorithm/Dijkstra/Algorithm.cpp
@@ -1,14 +1,5 @@
 ﻿#include "Algorithm.h"
 
-void Node::InputPosition(int height, int width)
-{
-
-	Position[0] = height;
-	Position[1] = width;
-
-}
-
-
 void Algorithm::NodeCheck(int* map[], int height, int width)
 {
 	int Direction = 0;
