+++ toda/Algorithm/Dijkstra/Dijkstra.vcxproj
@@ -126,9 +126,11 @@
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="Algorithm.cpp" />
     <ClCompile Include="Map.cpp" />
+    <ClCompile Include="Node.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Algorithm.h" />
+    <ClInclude Include="Node.h" />
     <ClInclude Include="Map.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
