+++ toda/Algorithm/Dijkstra/Node.cpp
@@ -1,0 +1,9 @@
+﻿#include "Node.h"
+
+void Node::InputPosition(int height, int width)
+{
+
+	Position[0] = height;
+	Position[1] = width;
+
+}
